Crate frtrs

A simple password safe, written in Rust.

Concepts

You must define a master password, which will be used to encrypt the vault. Each time you want to use the vault, this password will be asked you.

First, create a vault file:

frtrs create

Then, add entries to the vault (see docs to know more about the arguments):

frtrs add <identifier> --username <username> --password <password>

Security Principles

The master password is not stored

We are using only well-known methods and libraries

String checks and tests before releases

Independent code audit

Installation

Note: This project requires Rust 1.56+ to build.

Clone the repository

git clone https://github.com/xavier2p/fortress && cd fortress

Install the binary

cargo install --path .

Usage

$ frtrs --help
A simple password safe CLI app

Usage: frtrs [OPTIONS] [COMMAND]

Commands:
create  Create a new vault
copy    Copy the password of the desired identifier
view    View the password of the desired identifier
remove  Remove an entry from the vault
add     Add a new entry to the vault. If no one of the password methods is provided, the password will be the content of the clipboard
list    List all entries in the vault
help    Print this message or the help of the given subcommand(s)

Options:
-f, --file <PATH>      The input file path [default: /tmp/vault.frt]
--log-file <PATH>  Path to a file to write logs to [default: /tmp/fortress.log]
-h, --help             Print help
-V, --version          Print version

Modules

commands 🔒

The commands of the program.

crypto 🔒

Cryptographic functions

helpers 🔒

Utility functions and structs.

Functions

main 🔒

The main function, in which all magic happens.